Add image cors configuration
