+++ src/main/java/com/mealtiger/backend/BackendApplication.java
@@ -6,9 +6,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.Properties;
 
@@ -35,24 +32,6 @@ public class BackendApplication implements CommandLineRunner {
                     .build()
                     .run(args);
         }
-    }
-
-    @Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                Configurator configurator = new Configurator();
-                String allowedOrigins;
-
-                allowedOrigins = configurator.getString("REST.corsAllowedOrigins");
-
-                String[] allowedOriginsArray = allowedOrigins.split(",");
-
-                registry.addMapping("/recipes").allowedOrigins(allowedOriginsArray);
-                registry.addMapping("/recipes/**").allowedOrigins(allowedOriginsArray);
-            }
-        };
     }
 
     /**
